+++ src/helpers/helper.ts
@@ -1,0 +1,3 @@
+// For helpers that concern to a component, keep them in component as comp/helpers folder for easy navigation
+
+// For any functions that are independent and reusable of particular state/component etc
